[BUILDFIX] fix build for all architectures

+++ Source/DataConstructors/ShadeContourConstructor/ShadeContourConstructor.cpp
@@ -137,7 +137,8 @@ void ShadeContourConstructor::createContour(int idOfContour,
 	outputContour->poly = polyhedron;
 	SideOfContour* sides = outputContour->sides;
 	
-	DEBUG_PRINT("Number of visible edges = %d", edgesVisible.size());
+	DEBUG_PRINT("Number of visible edges = %ld",
+		(long unsigned int) edgesVisible.size());
 
 	EdgeSetIterator edgeCurr = edgesVisible.begin();
 	int iVertexCurr = edgeCurr->v1;
